requirements.txt
apiclient==1.0.4
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.8.0
bs4==0.0.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2==1.9.0.post1
oauth2-client==1.1.0
oauth2client==4.1.3
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pylint==2.3.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.16.0
wrapt==1.11.2
